Update q2.sh
Question 2:
Write a bash script that reads a file and counts the number of occurrences of each unique word, ignoring case sensitivity and excluding common English words such as "the," "and," and "in." Display the word count in descending order of frequency. You can create your own test files and share that with is when we submit your responses.